refactor: change relative imports to vite/ts path aliases (desc)
Relative paths are now used for files in the same scope as file which you import in. Imports which are more specific and doesn't include path such as "@auth" are considered shared and can be imported everywhere. Other imports like "@features/home/home" or "@layouts/default-layout/default-layout" are app scope, so they can be imported at the top level of app.tsx and app-routes.tsx, but not in shared modules. Currently enforcing module boundaries is turned off, but imports might be refactored to use this Nx feature in the future.